Support generating multiple declarations with the same name
Summary: It is not uncommon to have a function and a class with the same name.  When generating declarations from the set of dependencies, we put the declarations into an `SMap` keyed by name, and if a function and a class happen to share the name one of the declarations gets lost.

Reviewed By: Wilfred

Differential Revision: D17738063

fbshipit-source-id: 5ae6f2e5baf34660dce22ad609d2d6c8f6742ee7