Store proper closure ctx directly when inlining
Summary:
Instead of storing the closure object into ctx field, then loading it, grabbing
the closure context from it and storing it again, just store the correct thing
directly.

This restricts the use of InitCtx to prologue only. Removal of complexity
related to InitCtx when inlining is deferred to another diff.

Reviewed By: markw65

Differential Revision: D17725975

fbshipit-source-id: 376ec00ebb1adf82cac34c053509f311da64d5c7