Fetch class element types lazily
Summary:
It looks like {D17711777} unintentionally started eagerly fetching the types of all class elements when we fetch a class from the heap. This requires a separate round-trip to shared memory for each element (whereas previously we would skip doing this for all the elements we do not use), and it seems to have caused a perf regression in Zoncolan (see comments on D17711777).

We can avoid the eager fetch by making the new field `ce_pos` lazy (as `ce_type` is).

Reviewed By: fahndrich

Differential Revision: D17744322

fbshipit-source-id: c6a917ada5ecea2f921aab7b80bcdc9a6fad6510