Improve DecRef code-gen for mixed counted/uncounted type unions
Summary:
While emitting vasm for a dec-ref when we don't statically know if a
type is counted, we'll emit a runtime check, then emit the actual
dec-ref. For the release function call, we'll emit an indirect call
because we don't statically know the release function to
call. However, we're missing out on a source of type refinement
here. After the runtime countedness check, we know the type is
counted. This means we can potentially refine the type into one which
has a statically known release function.

For example, consider TStr | TInitNull. Once we've done the
countedness test, we know its just TStr and we can call
StringData::release.

Take advantage of this refinement using negativeCheckType(). We can't
just use type intersection with TCounted here, because that would turn
TStr into TCountedStr, and the string still might be static.

Reviewed By: markw65

Differential Revision: D17750219

fbshipit-source-id: af81001646e3349d968e576195833b7586d2a192