Respect the <<__Soft>> annotation even when no generics are passed
Summary:
`<<__Soft>>` annotation was supposed to mean that if for
```
function f<<<__Soft>> reify T>() {}
```
if we call this function `f` without this generic `T`, `f();`, it should raise a warning and assume it is erased instead of raises an exception. This worked fine when there was at least one erased generic but unfortunately did not work when no generics were provided. Lets fix this.

Ditto for class creation.

Reviewed By: markw65

Differential Revision: D17750442

fbshipit-source-id: 9f781a44fee690bb3f6dea88e0c185d96fc36a01