Clean rewrite of Hhbc_options in Rust
Summary:
Model emitter options (flags & arguments) in Rust
in a way that doesn't require so much boilerplate as in OCaml
(3-4 names/aliases + getter, setter, to_string, set_value, from_config).

On a high level, options with the same prefix (e.g., hhvm.hack.lang) are
grouped into a struct hierarchy so they are (de)serializable out of the box
via serde_json crate.

Most of the complication in the new code is handling discrepancies
between options names depending on the source:
- JSON config (preferred) - this is chosen as canonical name;
- CLI key=val (often without underscores and differently prefixed)

The former source is chosen as default because of unique names (bijection),
so JSON config names are automatically derived as follows:

   // (de)serializes field "hhvm.create_in_out_wrapper_functions"
   impl_flags!( HhvmFlags, "hhvm.", CREATE_IN_OUT_WRAPPER_FUNCTIONS, ...)

Deserializing from the latter (CLI) is achievable "for free" by:
- constructing and intermediate JSON representation (via `json!` macro); and
- merging it with the rest of JSON-deserialized config (former);

since those names are aliases in JSON serialization logic.

Most (>50) of the options are boolean, i.e., flags, so the above machinery
allows packing of flags into bits for efficiency (which proved to
be a bottleneck in parser several times), **instead of** boilerplate:

  #[serde(
    default,
    serialize_with = "serialize_flag_from_truish_string",
    deserialize_with = "deserialize_flags_from_truish_string",
    rename = "hhvm.create_in_out_wrapper_functions"
  )]
  pub create_in_out_wrapper_function: Arg<Option<bool>>;

Reviewed By: shiqicao

Differential Revision: D17495805

fbshipit-source-id: 1a5757bbcb59bbfd585ba7d8cdbff7bbd3cc73b8