Change serverSignaturehelp to use AAST directly
Summary:
When the parser is rustified, the legacy ast should hopefully disappear along with it. So I am trying to front run some of that work by removing existing references to the legacy by containing them to an implementation detail in lowering in full_fidelity_ast and removing calls that expose the legacy ast in the mli file. That way, when the lowerer is ported and goes straight to the aast, it can just be a drop in for everything.

This diff removes `lower_tree` and replaces it with an identical function that returns the aast.

Reviewed By: shiqicao

Differential Revision: D17695713

fbshipit-source-id: 041568b3a7dac71a88a383c13de86b50cc66d6ba