Delete hh_single_parse
Summary: When the parser is rustified, the legacy ast should hopefully disappear along with it. So I am trying to front run some of that work by removing existing references to the legacy by containing them to an implementation detail in lowering in full_fidelity_ast and removing calls that expose the legacy ast in the mli file. That way, when the lowerer is ported and goes straight to the aast, it can just be a drop in for everything. hh_single_parse seems to be one outstanding place where the legacy ast is exposed and used.

Reviewed By: losvald, vassilmladenov

Differential Revision: D17695711

fbshipit-source-id: cbd7182690014b5b187790aaba2d789b94b2f9c7