Add into_state to parsers.
Summary: The direct decl parser returns the decls using its state field, rather than its token type.  In order to move the state decls into an ocamlrep arena, they need to be consumable.  This change adds a new trait method to allow consuming the parser and returning an owned state.  I considered just using `.clone()`, but this feels cleaner because we truly are not using the parser afterward.

Reviewed By: ljw1004

Differential Revision: D17560125

fbshipit-source-id: 780efc22340e82c86c0e0c8d0301b5e9d1a5036a