Frame state track ctx through calls
Summary:
We noticed that DefInlineFP was not being elided due to LdCtx usage in the
inlined frame.  This prevented all sorts of optimizations we can do with the
dropped DefInlineFP.
Not dropping the ctx SSATmp in frame state across calls allows us to drop
LdCtx instructions in almost all cases.

There are a couple other changes that became necessary to pull this off.
Notably:
- Refcount opts needs to stop handling the context specially.  This exposed
  issues with throwing all loads of the context into a special must alias set.
  Fortunately there should only be one LdCtx instruction in a func, so we can
  just get rid of that logic.
- Tests that relied on us not eliding the inline frame pointer need to have
  some funcs marked as never inline.

Reviewed By: jano

Differential Revision: D17188454

fbshipit-source-id: a4f2f5f02723def53ceebf0547fce049d95f01af