introduce new type of types: constraint types
Summary:
We introduce a new type of types for types that can only live in the constraint graph: `contraint_type`. These types will be things like `HasMember(name, type)`, `IsArrayGettable(idx_type, type)` and so on.
We introduce the type of all locl types and constraint types: `internal_type` as `LoclType of locl ty | ConstraintType of constraint_type`.

We then change the type of the `upper_bounds` and `lower_bounds` field in `tyvar_info` to be sets of `internal_types` (`Internal_type_set` or `ITySet`), and change the subtype function to operate on internal types instead of just locl types.

For many functions in Typing_subtype, we now have a version that takes locl ty as argument, which will be used in most places outside of Typing_subtype, and a version that takes an internal type, whose named is suffixed by `_i` (e.g. `sub_type` and `sub_type_i`). The former calls the latter by wrapping its argument in the `LoclType` constructor.

`bind_to_upper/lower/equal_bounds` already makes sure we don't bind to constraint types.

This diff is still full of TODOs, but compiles. It is landable as we don't create any ConstraintType for now. I have diffs further up the stack which fix the TODOs.

Advantages:
- we can ensure constraint types don't "escape" from the constraint graph: tyvars can't be bound to a constraint type, and constraint types can't be used in the TAST, so can't be seen by users.
- we don't have to match on those new constraint types everywhere in the code, in places where they can't occur anyway at present.

Reviewed By: andrewjkennedy

Differential Revision: D16165375

fbshipit-source-id: 46bab364ed73e5312398a71f04bdecd146ca6de1