Move checkGC() to mallocSmallSizeSlow()
Summary:
In the same spirit as D16818473, reduce the number of slow paths by one,
by tucking the rare, but important, checkGC() into the slow path by making
the fast path fail when that check is needed.

On the fast path, we already decremented a counter and branched when it
went negative; this part is unchanged, but now we branch to a single shared
slow path function.

Reviewed By: ricklavoie

Differential Revision: D17070289

fbshipit-source-id: dc7e03474d7a0111039994b41d5da2cc87c74115