Fix a docblock contents issue
Summary:
While researching some HHI documentation tests, I discovered that the "option.value_exn" that gathered file contents in serverless IDE was sometimes throwing an exception.  I believe this was being caused by the usage of the "cleansed HHI path" value.

This diff enables the code to work even if the file contents value is set to None.  It also corrects the cleansed HHI value and adds a comment so I don't accidentally change it again in the future.

Reviewed By: arxanas

Differential Revision: D17794144

fbshipit-source-id: cd3706f4d49f88a6e7387b08b1dcea64635d1cbd