Enable tests for autocomplete in sqlite and serverless IDE
Summary:
In order to deliver a better experience overall, we should have tests running using the sqlite provider (which is most similar to the custom provider used inhouse).

Now that we've matched the behavior of sqlite to the autocomplete sharedmem/trie as closely as possible, let's switch all tests to use sqlite specifically.

Some notes about expectations:
* The "Data" section of LSP results changes depending on our provider.  A lot of the lines of changes in this diff are from slightly different "data" sections, which is invisible to user experience.  The majority of these are because if we have a full symbol name we no longer need filename / line / column.

Reviewed By: arxanas

Differential Revision: D17796727

fbshipit-source-id: 3a45ac3eea99fdbaee4dc6b2b41d90b703b7dc3e