Fix incorrect "complete" flag values
Summary:
As discovered in a meeting today, there are some rare cases where two separate symbols with the same text will appear in search results.  There was an old "dedup" command in the symbol index pathway; when that command detected two identical symbols (which might have differing kinds), it would incorrectly remove them.

Let's fix that, and also take this opportunity to correct a small error in computing when we'd reached the right limit of results for autocomplete by measuring the ref'd variable rather than the temporary local one.

Reviewed By: arxanas

Differential Revision: D17800261

fbshipit-source-id: eaa8346841b0d574a0a18febfbbc85dc5c8e6706