Put declarations originating from partial mode files into partial mode sections of the output
Summary: At the moment, the output produced by `hh --extract-standalone` is a possibly multi-section file containing at most two strict mode sections.  However, some of the dependencies we are pulling in might be defined in partial mode files (e.g., a function might be lacking the return type, or an instantiation of some generic type might be lacking generic type parameters).  Such definitions will not type check if they are placed into a strict mode section.  Therefore, we need to split the declarations based on the mode of the file from which they originate and place them into appropriate mode sections of the output.

Reviewed By: Wilfred

Differential Revision: D17764031

fbshipit-source-id: 362a6859f519088297090d733996639895ae06a6