Delete TempContinue and TempBreak
Summary: `TempContinue` and `TempBreak` were introduced to represent numbered `break` and `continue` (e.g. `break 1`, `break $foo` where `$foo` is an integer, etc.) on the AAST. This was needed to ensure all code continued to work when we migrated the bytecode emitter from using the legacy AST to using the AAST. However, we don't need these constructors anymore so this diff deletes them and updates/deletes dependent files & tests.

Reviewed By: Wilfred

Differential Revision: D17202552

fbshipit-source-id: 32b99991da806e9caebef6f5bb4ab19c27800baf