Check slab integrity
Summary:
Since cachelib only exposes byte sequences to us, we do not have the ability to use the type system to guarantee invariants about slabs. Since attempts to use an invalid or corrupt slab would result in panics or segfaults, it would be useful for us to sanity-check the contents of a slab before attempting to use it.

In the previous diff, we use a magic number as a coarse means of doing this sanity-check, but there many other invariants we want to uphold:

- The size of the slab is large enough that metadata accesses will not panic (with out-of-bounds)
- All pointers point to a block value (that is, one word after a block header) within the slab's block-data region (or would do, after being corrected in a rebase)
- The size in each block header specifies a range of addresses which lies entirely within the block-data region
- The root value offset points to a block value in the block data region

This might be overly-paranoid in practice, and might involve a runtime cost we are not willing to pay in the final version of the decl service, but I expect that even if that is the case, it will be helpful as a sanity check during development.

Reviewed By: jewelpit

Differential Revision: D17491481

fbshipit-source-id: 34f18b81a2a50bc2e8ae69b7fdebe9e6db08af34