Do not allocate duplicates of values at the same memory address
Summary:
There's a lot of repeated values in decls (filenames in particular) which do not occupy redundant space in the OCaml heap because they are shared. Currently, ocamlrep does not maintain this sharing--it makes a copy of every shared value tree for each reference to it.

While maintaining sharing in Arena might be a bit more involved, it is very simple to maintain this sharing when recursively copying values using SlabBuilder::copy_value. We just need to keep track of the pointers we've seen and the locations of their copies in the slab.

Reviewed By: jewelpit

Differential Revision: D17491744

fbshipit-source-id: 2e2e9110f980bf165b544d22a9d9afb2cc6c0b03