Store entire language status request in memory
Summary: We currently store some data that we can use to derive the language status request in the language server's state. However, when we add sIDE status, we need to merge more data than just that (such as progress indicators). Instead, store the raw LSP request, so that we have access to all the fields we need to merge together.

Differential Revision: D17585762

fbshipit-source-id: 7b794a943b97f9bbddcd97e7385288686de3de30