Back out "Serverless IDE startup message"
Summary: Original commit changeset: 5970741c50ed ({D17574073}). We landed that diff to ensure that a version of sIDE with status was shipped, but the easiest way to resolve the merge conflicts is to back it out. The following stack includes generalizations of the contents of this diff and I didn't want to attempt to split this commit up across the appropriate following commits. (For example, instead of `is_running`, we have `get_status`, which returns a more descriptive status about the IDE service.)

Differential Revision: D17633888

fbshipit-source-id: 16e13df5f557fa0cffc9312b1fdcf23962f97bfe