Favor standard elaborated identifier in even more cases
Summary:
This is a behavior change to make HH-autoimporting of classes consistent, both between the typechecker and compiler AND between classes and functions/consts in the compiler. After this diff, any identifier prefixed with a backslash *always* indicates a fully-qualified name.

```
namespace Foo;

// Before:
new \Vector(); // Refers to HH\Vector in compiler, \Vector in typechecker

// After:
new \Vector(); // Refers to \Vector in compiler and typechecker
```

This also fixes a bug in which the specific identifier "namespace" would always be elaborated into the current namespace, but I'm not even sure that's legal.

Differential Revision: D17814455

fbshipit-source-id: 84a232657d03e4303f72fdde07938a56f4a53029