Strip HH\ prefix for autoimported names in Typing_print
Summary:
This updates Typing_print to strip the `HH\` from autoimported names.

We do this because some extremely common names are autoimported, like Collections and Awaitables. As we unify the typechecker and compiler's understanding of autoimported types, we don't want the `HH\` to appear in error messages and hover types, because it may confuse users into writing that in code. This has already started happening for the uncommon names: https://fburl.com/codesearch/47r29lye

For example, D15693424 unified autoimport behavior for type structures, and `HH\` appeared in the error messages involving TypeStructureKind. This diff reverses that change.

Differential Revision: D17694748

fbshipit-source-id: 165dc8bebaaeba22772e1e8461fa55f3a022b159