When adding a dependency on an enum, also add dependencies on all its values
Summary:
Consider the following code:
```
enum E: int {
  A = 1;
  B = 2;
}

function foo(E $e): ?int {
  switch ($e) {
    case E::A:
      return $e;
    default:
      return null;
  }
}
```
At the moment, extracting `foo` from this code will pull in the dependency on `E` and on its value `E::A` but not `E::B` resulting in the following extracted code:
```
enum E: int {
  A = 0;
}

function foo(E $e): ?int {
  switch ($e) {
    case E::A:
      return $e;
    default:
      return null;
  }
}
```
This code produces a type error because the `switch` statement covers all the cases and the `default` case is redundant.

To avoid this situation, let us pull in all the values of the enum as dependencies.

Reviewed By: andrewjkennedy

Differential Revision: D17830035

fbshipit-source-id: 25019c43457475d59579eac165bf953e805efe3e