Remove duplicate name re-use check from emitter
Summary: Checking that users don't `use` a name twice is done at the parser stage, so we don't need this check at the emitter stage. Removing this check is necessary for D17817943 to work, because unlike `use` statements written in code, we want the implicit `use` statements to be overridable.

Differential Revision: D17824626

fbshipit-source-id: 655fa0e59f688ddd7d04433f5aa40cb31ba550aa