Persist inferred type arguments in TAST
Summary:
TAST expressions are annotated with their inferred types. But the types that Hack infers for generic arguments (to constructors, functions and methods) are not present in the TAST. Lint rules, for example, must somehow reconstruct them, and it's hard to implement a type checker for the TAST without this information.

So let's persist inferred type arguments. We do this by annotating *type argument hints* with localized types. If the hint is explicit (e.g. `foo<int>(3)`), this is redundant, but useful (because it has been localized). If the hint is a wildcard (e.g. `foo<_>(3)`), then the type that Hack filled in for the wildcard is now present in the TAST. If no hints are present at all (e.g. `foo(3)`) then wildcard hints are included in the TAST. This means that as far as the TAST is concerned, there is no distinction beteween `foo(3)` and `foo<_>(3)`, but the latter should be considered sugar for the former anyway.

Reviewed By: Wilfred

Differential Revision: D17449923

fbshipit-source-id: 74f8255ae6eca2137152a42295deac607583587f