Kill from_file and replace with from_file_with_legacy
Summary:
When the parser is rustified, the legacy ast should hopefully disappear along with it. So I am trying to front run some of that work by removing existing references to the legacy by containing them to an implementation detail in lowering in full_fidelity_ast and removing calls that expose the legacy ast in the mli file. That way, when the lowerer is ported and goes straight to the aast, it can just be a drop in for everything.

This diff kills the final interface function from full_fidelity_ast that returns a legacy ast. Everyone that previously consumed a lowered legacy ast now consumes a new aast and it should be straightforward to drop in the aast without any difficulties.

Reviewed By: shiqicao

Differential Revision: D17696334

fbshipit-source-id: 93066e7e93ec4a265355a8032bc0abd3b52817e6