Refactor make_all_members
Summary:
`make_all_members` provides a function that creates a `Dep.variant` for a given class and member name. This is confusing because it's only used by one of the functions (`check_members`) and the others throw it away.

Move the `Dep.variant` creation to `check_members` and simplify the other functions.

Differential Revision: D17834555

fbshipit-source-id: 4187019b153e8cbc72381906b896efec4d2a67ed