Simplify linter_equality_check by using inferred type arguments
Summary:
Now that the TAST records type arguments to generic functions, it's easier to implement type-dependent lint checks by looking at the type arguments, as here. Note that the signature of `contains` and `contains_key` now use different type parameters for the value and the element type (or key type, for `contains_key`) of the traversable. The lint rule now inspects the type arguments instantiated for these parameters, removing the need for messy inspection and subtyping against the `Traversable` type.

The tests now need a preamble that defines `contains` and `contains_key` functions.

Reviewed By: manzyuk

Differential Revision: D17842711

fbshipit-source-id: 7a9eec8438d67599c1f1794f5db53a0081ab0d5d