Constants for sharing binary data in tracing
Summary:
Strobelight and HHVM will be communicating stack data via
BPF. This is done via slabs of memory with fixed width fields. This
lib defines the sizes and numbers of these fields.

The values here are conservative guesses which should be more data
than we need in almost all cases, and are negotiable.

Reviewed By: billf

Differential Revision: D17631166

fbshipit-source-id: 20b859d8628e2b1337324873f4ec59abeadb6ad0