Show pessimistic language status
Summary:
If either serverless IDE or `hh_server` is not working, show an error language status. (Or similarly, if either is processing, show the warning language status.)

Facebook
ljw1004 thinks that we should primarily indicate whether the IDE is functioning. But at least for now, I believe that users are expecting errors for the entire codebase as part of the IDE functioning. Even if we change their workflow to explicitly request errors, they should at least know when they have to do so.

Note that VS Code currently doesn't display language status at all if it's info-level.

Another possible solution would be to surface a dummy diagnostic in the "Problems" window indicating that diagnostics are out of date, or use the ability to tie diagnostics to a version of the document (to be introduced in LSP 3.15, hopefully with some kind of editor support).

Differential Revision: D17801062

fbshipit-source-id: 0b9f066eee18849270a755bb8715350464e2dbf8