Avoid global emitter sideeffect from dispatch_loop
Summary:
Most flags passed to emitter are deserialized from JSON
right before the emitter is called (from header and/or message),
but the `is_systemlib` flag is read from the dispatch_loop
and passed to "global emitter state" right there.

This poses a problem for migrating to Rust because emitter
state doesn't exist before the "code" handler is invoked
(even if it existed, emitter's lifetime would be unpredictable).

Reviewed By: shiqicao

Differential Revision: D17862865

fbshipit-source-id: 48c9884617fad17e63a0e318e5ace794f5153448