Add a debug tool to skip function and method bodies
Summary: The code that handles timeouts isn't always effective in terminating typechecking of a method body, causing the typechecker to hang. This creates an attribute similar to `// UNSAFE` that simply replaces the method body with `[]` and throws an internal error. This can help selectively disable expensive functions during development. For methods, the lack of body will throw a NastCheck error about the method needing to be abstract, but this is not a concern.

Reviewed By: Matt-Schellhas

Differential Revision: D17863164

fbshipit-source-id: 1ac6e3e04f7669fd49efe99b9ab03048db6a80d0