first step towards typechecking
Summary:
More baby steps for hh_mapreduce. Here it's setting up environment to allow typechecking.

(It's not complete yet. I only did it far enough to be able to compile. It still requires logging setup, for instance).

I feel uneasy about this environment setup. I'm not sure I've set all the necessary global variables. I kind of want to (1) audit every single of the thing that's set up and figure out where every setting came from and why it's needed, (2) figure out what parts should be done by the decl service, which parts should be done by the mapreduce orchestrator, and which parts should be done by the worker.

Anyway, there's lots to be done here, so this is just the first steps.

Reviewed By: arxanas

Differential Revision: D17811432

fbshipit-source-id: 41736eff0be464cf7d8344c985fa8208ce63c4f3