a provider for decl-service
Summary:
A "provider" is a generic concept that encompases how ASTs will be provided to the server, how file-contents will be provided to the server, how decls will be provided to the typechecker, how fixmes will be provided to the typechecker, how parser-options will be provided to the parser.

Waleed already made a "local memory provider" with a few hand-carved cases for what sIDE will want provided.

Here I'm doing the same kind of thing for what provisions there'll be when we use the new architecture (i.e. no shared-mem, and a decl service).

This diff merely introduces the new provider kind, in provider_config.ml. The rest of the changes are sticking in "not yet implemented" in a bunch of places that match on provider.

Reviewed By: arxanas

Differential Revision: D17826505

fbshipit-source-id: 9d4346f867537619b24c3bdf2191adb32e926020