accessing C::T when T is abstract should be forbidden in some cases
Summary:
Writing the type hint C::T should not be allowed if T is abstract in C.

For example, the followings don't make sense:
```
abstract class C {
  abstract const type T;
}

function foo(C::T $x): void {}
function bar(): C::T { ... }
class A extends B<C::T> { ... }
```
and so on.

In this diff:
In Typing_taccess, where we resolve type constant accesses, if we resolve to an abstract type constant, this is an error, except if the lhs of `::` is anything Tabstract.

Unfortunately, `this` and generic type parameters are sometimes localized to a concrete class so we need to detect them in typing_phase and pass a flag to `expand_typeconst` accordingly.

Besides, since `localize`, which calls `expand_typeconst`, is called all over the place, the error is often duplicated, and sometimes wrapped in other errors, e.g. when checking class extensions and implementations, producing confusing errors. To avoid that, a new flag is added to `expand_env`. It's quite unsatisfactory, but T54121530 aims at finding a better solution.

Reviewed By: manzyuk

Differential Revision: D14186596

fbshipit-source-id: ce45266f94c4268b0b883f2dde338b9dcf4eb5f4