Finish killing hphp/runtime/base/externals.h
Summary:
None of the callers of invoke pass false for the third arg, so remove it. That arg was only passed through to the invoke_failed helper, but when true is passed to that helper it is reduced to a single throw; inline the throw into invoke.

All the callers of invoke have the function as a String or std::string, so remove the wrapper that takes a const char* and turns it into a const String& and call the latter directly.

Move the declaration of invoke into builtin-functions.h. Kill externals.h.

Reviewed By: billf

Differential Revision: D17873459

fbshipit-source-id: 6980e1a516ca321bb192fb31e783f1d4f739dda1