Track and log ArrayIterProfile
Summary:
We'll specialize certain iterators based on profiling. Our new profile tracks:
1. The layout of the base: PackedArray, MixedArray, Vec, or Dict
2. The types of the keys in the base. It's just ints for the vec-like arrays, but we check MixedArrayKeys for the dict-like arrays.
3. The types of the value types. Since we only profile once (at the init), we iterate over multiple values - the first 16.

We'll make use of all of this information when generating specialized code:
- We need a consistent base type for specialization to be worth it - otherwise, we'll have conditionals everywhere we access or increment the position.
- For dict-like arrays, we'll guard the specialized IterInit behind an O(1) type-check against the key bit-set, then assert the key type when loading it.
- Whenever we access the array, we'll do a check against the profiled value type to eliminate other ref-count and check instructions.

In this diff, we also introduce the "IteratorSpecialization" byte that we'll set in the ArrayIterator for specialized iterators. This byte has enough information to guard against items 1 and 2 (the type of the base and the type of the keys); thus, we can avoid guarding on those things at the IterNext op if we check this one byte. Actually, the situation is even better - we can use memory effects to propagate information about the "IteratorSpecialization" byte and completely elide the IterNext check. In ServiceLab runs, I've never seen the check on this byte survive load-elim. The reason this byte is part of this diff is because it's the output of the new profile.

Reviewed By: mofarrell

Differential Revision: D16144713

fbshipit-source-id: 6a438f87accfa82aef5a479efbb2812bb6916227