dummy naming provider
Summary:
This diff provides a sort of dummy naming provider so we can typecheck files which include references to symbols in other files. The only kinds of symbols it supports so far are global constants in other files.

I think this area will need a refactor before proceeding with decl-service. The way it works is we do a "naming" phase. This looks up the naming table to get a FileInfo.t and then throws it away. It only uses the fact of whether or not it got an result from the naming table, in order to decide whether to report an "unbound name" error.

That might have made sense in the past but we're moving to a model where the decl service is the one responsible for giving out decls. It will have a single API "give me the decl for Foo", and it might return either "here's the decl" or "that symbol is undefined". We don't want a chatty conversation where it asks one question "is Foo defined" and then another question "give me the definition of Foo".

I don't know if there are problems to expect if we move the "unbound" checks out of naming and into typechecking? As far as I can tell, extracting shallow decls is wholly independent of a naming-phase having reported unbound name errors.

Anyway, this diff is just a placeholder so we can get to work on the rest of the skeleton.

Reviewed By: arxanas

Differential Revision: D17826875

fbshipit-source-id: f65b236d1c52173227b166dc1e82bdee6ed60ff2