dummy decl_provider and fixme_provider for decl service
Summary:
I wanted to make the mapreduce typechecker code work end to end. To achieve that, this diff provides...

A dummy `Decl_provider.get_gconst` (which is called when the typechecker needs to fetch the decl of a global const, and the dummy implementation here returns the answer that it wasn't found).

A dummy `Fixme_provider.get_fixmes_for_pos` (which is called when the typechecker found an error and has to establish whether the error is suppressed, and the dummy implementation here returns that there were no suppressions). Also a dummy `Fixme_provider.is_disallowed` (which I'm not entirely sure what it's for, but the dummy implementation here returns None).

These two changes are enough to typecheck a simple file that looks like the one in the test plan. I think the only thing that lets it scrape by is that there are no lookups of any decls from any other files.

It was at first a surprise to me that I needed to provide a dummy `Decl_provider.get_gconst`. I'd have thought that, since we're not looking up decls from any other files, we shouldn't need a decl. But what happens in typing.ml is `let tast = List.map nast convert_def`, which calls into `gconst_def`, which calls into `Env.get_gconst`. I guess it's not surprising that producing a TAST involves using decls to provide type hints. And of course those decls are fetched from the decl_provider, which is the only vehicle there exists for providing decls.

Reviewed By: arxanas

Differential Revision: D17826668

fbshipit-source-id: 5b65d45599d1382185ca3102287ee8d4022b7870