Add dependencies on type constants declared in classes / interfaces and inherited by traits through `require extends` / `require implements`
Summary:
Type constants cannot be declared in traits but can be inherited through `require extends` or `require implements`.  If we have a trait that references an inherited type constant, e.g.:
```
interface I {
  abstract const type T;
}
trait U {
  require implements I;
  public function get(): ?this::T {
    return null;
  }
}
```
then `this::T` gets expanded into `U::T`, so that we don't depend on `I::T` directly and hence when a declaration of `I` is generated in the extracted code, it doesn't have a declaration of the type constant `T`, which leads to a type error.

Reviewed By: kmeht

Differential Revision: D17906548

fbshipit-source-id: 1df98dd137007197d7987f7288565f2b95f13c84